🚚: moved `HistoryDatasetPicker` from the `libraries` to the `SelectionDialog` directory and updated import paths accordingly
